Extract sharing tests
